rename and split pg_upgrade tests
Make the pg_upgrade tests much more focused, clear, and precise.
