+++ test/acceptance/pg_upgrade_non_upradeable_test.go
@@ -1,0 +1,40 @@
+// Copyright (c) 2017-2023 VMware, Inc. or its affiliates
+// SPDX-License-Identifier: Apache-2.0
+
+package gpupgrade_test
+
+import (
+	"path/filepath"
+	"testing"
+
+	"github.com/greenplum-db/gpupgrade/idl"
+	"github.com/greenplum-db/gpupgrade/testutils"
+)
+
+func Test_PgUpgrade_NonUpgradeable_Tests(t *testing.T) {
+	// Since finalize archives the stateDir (GPUPGRADE_HOME) backups and
+	// migration scripts cannot be stored here.
+	stateDir := testutils.GetTempDir(t, "stateDir")
+	defer testutils.MustRemoveAll(t, stateDir)
+
+	resetEnv := testutils.SetEnv(t, "GPUPGRADE_HOME", stateDir)
+	defer resetEnv()
+
+	source := GetSourceCluster(t)
+	dir := "6-to-7"
+	if source.Version.Major == 5 {
+		dir = "5-to-6"
+	}
+
+	testDir := filepath.Join(MustGetRepoRoot(t), "test", "acceptance", dir)
+	testutils.MustApplySQLFile(t, GPHOME_SOURCE, PGPORT, filepath.Join(testDir, "setup_globals.sql"))
+	defer testutils.MustApplySQLFile(t, GPHOME_SOURCE, PGPORT, filepath.Join(testDir, "teardown_globals.sql"))
+
+	t.Run("pg_upgrade --check detects non-upgradeable objects", func(t *testing.T) {
+		nonUpgradeableTestDir := filepath.Join(testDir, "non_upgradeable_tests")
+		isolation2_regress(t, source.Version, GPHOME_SOURCE, PGPORT, nonUpgradeableTestDir, nonUpgradeableTestDir, idl.Schedule_non_upgradeable_schedule)
+
+		revert(t)
+	})
+
+}
